[FIX]: 피드백 반영, RMSNorm dtype mismatch + 노트북 perplexity overflow 2건
- src/graphlm/neuron/rms_norm.py — RMSNorm.forward 에서 self.weight 를 x_dtype 으로 cast
  (gemini #3303153077): mixed precision (FP16/BF16) 입력 시 weight 가 float32 로 남아 있어 출력 dtype 이 강제로 float32 promotion 되던 문제 — residual connection dtype mismatch 회피
- notebooks/02-function-level/12-phase13-hybrid-transformer.ipynb — safe_perplexity 헬퍼 도입
  (gemini #3303153101): 학습 초반 큰 loss 에서 math.exp OverflowError 회피 (cap=20.0)

+++ src/graphlm/neuron/rms_norm.py
@@ -49,7 +49,9 @@ class RMSNorm(nn.Module):
         x_f = x.float()
         rms = torch.rsqrt(x_f.pow(2).mean(dim=-1, keepdim=True) + self.eps)
         y = (x_f * rms).to(x_dtype)
-        return y * self.weight
+        # weight 도 x_dtype 로 cast — 그래야 mixed precision (FP16/BF16) 에서 residual
+        # connection dtype mismatch 회피 (gemini #3303153077)
+        return y * self.weight.to(x_dtype)
 
     def extra_repr(self) -> str:
         return f"hidden_dim={self.hidden_dim}, eps={self.eps}"
